requirements.txt
altgraph==0.17.5
bcrypt==5.0.0
charset-normalizer==3.4.4
et_xmlfile==2.0.0
greenlet==3.3.2
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pefile==2024.8.26
pillow==12.1.1
pip==26.0.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
python-dateutil==2.9.0.post0
pywin32==311
pywin32-ctypes==0.2.3
reportlab==4.4.10
setuptools==82.0.0
shiboken6==6.10.2
six==1.17.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
tzdata==2025.3
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-dotenv==1.0.1
psycopg2-binary==2.9.11
